Hygeine: Refactor tests to integration tests
